cast cycle count to int for logging

+++ RadioModule.cpp
@@ -210,7 +210,7 @@ void RadioModule::handleExtendedTransmitStatus(const uint8_t *frame, uint8_t len
     json.insert("RetryCount", status->retryCount);
     json.insert("DeliveryStatus", status->deliveryStatus);
     json.insert("Discovery", status->discovery);
-    json.insert("CycleCount", cycleCountsFromFrameID[status->frameID]);
+    json.insert("CycleCount", (int)cycleCountsFromFrameID[status->frameID]);
 
     log("Transmit status for frame ID %03x: %02x", status->frameID, status->deliveryStatus);
